+++ KPI-ejb/src/java/vn/hanbell/kpi/mail/BscGroupVNShipmentMailBean.java
@@ -7,7 +7,6 @@ package vn.hanbell.kpi.mail;
 
 import cn.hanbell.kpi.comm.MailNotification;
 import cn.hanbell.kpi.ejb.erp.BscGroupVHSaleOrderBean;
-import cn.hanbell.kpi.ejb.erp.BscGroupVHServiceBean;
 import cn.hanbell.kpi.ejb.erp.BscGroupVHShipmentBean;
 import cn.hanbell.kpi.entity.Indicator;
 import java.util.Date;
@@ -28,8 +27,6 @@ public class BscGroupVNShipmentMailBean extends MailNotification {
     private BscGroupVHShipmentBean bscGroupVHShipmentBean;
     @EJB
     private BscGroupVHSaleOrderBean bscGroupVHSaleOrderBean;
-    @EJB
-    private BscGroupVHServiceBean bscGroupVHServiceBean;
     
     public BscGroupVNShipmentMailBean() {
 
@@ -57,12 +54,12 @@ public class BscGroupVNShipmentMailBean extends MailNotification {
             //越南出货
             bscGroupVHShipmentBean.updataShpimentActualValue(y, m, d);
              log4j.info("End Execute Job updataShpimentActualValue");
-             //越南服务
-             bscGroupVHServiceBean.updataServerActualValue(y, m, d);
-             log4j.info("End Execute Job updataServerActualValue");
              //越南订单
              bscGroupVHSaleOrderBean.updataSaleOrderActualValue(y, m, d);
             log4j.info("End Execute Job updataSaleOrderActualValue");
+             //越南服务
+             bscGroupVHShipmentBean.updataServerActualValue(y, m, d);
+             log4j.info("End Execute Job updataServerActualValue");
             return "越南数据更新集团报表数据成功";
         } catch (Exception ex) {
             return ex.toString();
